feat(Path): 연령별 추가 요금 계산 시 로그인 사용자 구분

+++ src/main/java/nextstep/subway/path/domain/Fare.java
@@ -51,7 +51,7 @@ public class Fare implements Comparable<Fare> {
 
     public Fare add(final Fare o) {
         final BigDecimal newFare = getFare().add(o.getFare());
-        return new Fare(newFare);
+        return new Fare(BigDecimal.valueOf(newFare.longValue()));
     }
 
     public BigDecimal getFare() {
